requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-filter==2.4.0
django-oauth-toolkit==1.3.3
django-rest-framework==0.1.0
djangorestframework==3.12.2
drf-yasg==1.20.0
flake8==3.8.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.8
Pillow==8.1.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
python-decouple==3.4
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
